+++ OpenGPSLogger/app/src/main/res/layout/fragment_home.xml
@@ -81,7 +81,7 @@
             android:layout_weight="1"
             app:layout_constraintTop_toBottomOf="@+id/progress_bars" >
 
-            <eu.tijlb.opengpslogger.ImageRendererView
+            <eu.tijlb.opengpslogger.view.ImageRendererView
                 android:id="@+id/imageRendererView"
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
